requirements.txt
Flask==2.3.2
google-api-python-client==2.99.0
google-auth==2.23.0
google-cloud-firestore==2.15.0
requests==2.31.0 